PT-2707: Moved remove/add members methods and setPedigree to PatientsInFamilyManager.
