+++ components/family-studies/api/src/main/java/org/phenotips/studies/family/internal/PhenotipsFamilyTools.java
@@ -25,6 +25,7 @@ import org.phenotips.studies.family.FamilyRepository;
 import org.phenotips.studies.family.FamilyTools;
 import org.phenotips.studies.family.Pedigree;
 import org.phenotips.studies.family.exceptions.PTException;
+import org.phenotips.studies.family.groupManagers.PatientsInFamilyManager;
 
 import org.xwiki.component.annotation.Component;
 import org.xwiki.model.EntityType;
@@ -49,6 +50,9 @@ import javax.inject.Singleton;
 @Singleton
 public class PhenotipsFamilyTools implements FamilyTools
 {
+    @Inject
+    private PatientsInFamilyManager pifManager;
+
     @Inject
     private FamilyRepository familyRepository;
 
@@ -155,7 +159,7 @@ public class PhenotipsFamilyTools implements FamilyTools
         }
 
         try {
-            this.familyRepository.removeMember(family, patient, currentUser);
+            this.pifManager.removeMember(family, patient);
         } catch (PTException ex) {
             return false;
         }
@@ -177,7 +181,7 @@ public class PhenotipsFamilyTools implements FamilyTools
     @Override
     public boolean forceRemoveAllMembers(Family family)
     {
-        return this.familyRepository.forceRemoveAllMembers(family, this.userManager.getCurrentUser());
+        return this.pifManager.forceRemoveAllMembers(family);
     }
 
     @Override
@@ -222,7 +226,7 @@ public class PhenotipsFamilyTools implements FamilyTools
     @Override
     public void setPedigree(Family family, Pedigree pedigree) throws PTException
     {
-        this.familyRepository.setPedigree(family, pedigree, this.userManager.getCurrentUser());
+        this.pifManager.setPedigree(family, pedigree);
     }
 
     @Override
